+++ Changelog.md
@@ -81,3 +81,8 @@ Continued working on the [Story1](https://github.com/Ilovemyhous/Who/blob/main/G
 *9th June 2022*
 The README file was a bit updated.
 Continued working on [Story1](https://github.com/Ilovemyhous/Who/blob/main/Game/Story1/HTML.html), almost done.
+
+## Version 0.0.16
+*14th Junes 2022*
+Finalised the [Story1](https://github.com/Ilovemyhous/Who/blob/main/Game/Story1/HTML.html).
+Started the puzzle 3.